pyproject.toml
[project]
name = "seds536-skin-tone"
version = "0.1.0"
description = "Skin Tone Detection for Skincare Recommendations - SEDS536 Term Project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core ML
    "torch>=2.0.0",
    "torchvision>=0.15.0",

    # Data handling
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pillow>=9.0.0",

    # Image processing
    "opencv-python>=4.8.0",

    # Configuration
    "pyyaml>=6.0",

    # Metrics and evaluation
    "scikit-learn>=1.3.0",

    # Progress bars
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
conversion = [
    "ai-edge-torch>=0.3.0",
    "tensorflow>=2.15.0",
]

[project.scripts]
train = "training.scripts.train:main"
evaluate = "training.scripts.evaluate:main"
check-env = "training.scripts.check_env:main"
preprocess = "training.scripts.preprocess_ccv2:main"
preprocess-faces = "training.scripts.preprocess_faces:main"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

# Use PyTorch CUDA 12.8 index for GPU support
[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["training"]
